refactor(styled-react): update how we re-export components
