+++ packages/react/src/index.ts
@@ -151,6 +151,7 @@ export type {StateLabelProps} from './StateLabel'
 export {default as SubNav} from './SubNav'
 export type {SubNavProps, SubNavLinkProps, SubNavLinksProps} from './SubNav'
 export {default as ToggleSwitch} from './ToggleSwitch'
+export type {ToggleSwitchProps} from './ToggleSwitch'
 export {default as TextInput} from './TextInput'
 export type {TextInputProps} from './TextInput'
 export {default as TextInputWithTokens} from './TextInputWithTokens'
